requirements.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
channels[daphne]
shortuuid
pygame
